requirements.txt
Flask==3.0.0
numpy==1.26.1
Pillow==10.1.0
Pillow==10.1.0
pywebview==4.3.3
scikit_learn==1.3.1
